Files are now saved in a temporary local folder.
Previously a local cache was used so that files that had previously
been downloaded from Google Drive didn't need to be downloaded again.
However, following an investigation into the data storage, it appears
that files in Google Drive are continuously being written to and cannot
be regarded as final copies.
Rather than hashing these files to form a local cache, the program now
just redownloads all the available files.

As the amount of data increases this behaviour might need to be changed
to use a caching system again.