Downloads met data too (#1)
* Downloads met data too

* Empty values are removed

* Maps compass points to numerical degrees

* Met data pulled from raw binary file

Previously met data was read from parsed CSV files. It is now read from
the binary .wlk files. This places the need parsing need from data
collection to data analysis, which makes more sense.

* Output a wide CSV rather than long

This should result in quicker computations in the web-app, as it won't
need to convert between long and wide